requirements.txt
beautifulsoup4==4.12.3
mysql-connector-python==8.3.0
selenium==4.18.1
webdriver-manager==4.0.1
